+++ 1.20.1/Forge/src/main/java/fuzs/spikyspikes/SpikySpikesForge.java
@@ -33,6 +33,7 @@ public class SpikySpikesForge {
         final ExistingFileHelper fileHelper = evt.getExistingFileHelper();
         dataGenerator.addProvider(true, new ModBlockLootProvider(packOutput, SpikySpikes.MOD_ID));
         dataGenerator.addProvider(true, new ModBlockTagsProvider(packOutput, lookupProvider, SpikySpikes.MOD_ID, fileHelper));
+        dataGenerator.addProvider(true, new ModEntityTypeTagProvider(packOutput, fileHelper, SpikySpikes.MOD_ID, lookupProvider));
         dataGenerator.addProvider(true, new ModDamageTypeProvider(packOutput, SpikySpikes.MOD_ID, fileHelper));
         dataGenerator.addProvider(true, new ModLanguageProvider(packOutput, SpikySpikes.MOD_ID));
         dataGenerator.addProvider(true, new ModModelProvider(packOutput, SpikySpikes.MOD_ID, fileHelper));