+++ 1.21.1/CHANGELOG.md
@@ -1,12 +1,22 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.1.1-1.21.1] - 2025-08-25
+
+### Added
+
+- Add `spikypikes:spike_damage_immune` entity type tag
+
 ## [v21.1.1-1.21.1] - 2025-05-01
+
 ### Changed
-- Improve mod compatibility for spikes that yield experience drops 
+
+- Improve mod compatibility for spikes that yield experience drops
 
 ## [v21.1.0-1.21.1] - 2024-09-17
+
 - Port to Minecraft 1.21.1